fix(runtime): refuse Node declarations the resolver cannot trust
- runtime.node as a decimal YAML number (22.10 loads as 22.1) now errors and
  asks for a quoted version, in the resolver and in both schemas.
- A version holding whitespace, a non-string package.json Node field, a
  non-object package.json, or a runtime that is not a mapping now errors
  instead of defaulting or corrupting GITHUB_OUTPUT.
- A trailing # comment in .nvmrc/.node-version is dropped.
- The source names the config file actually read (CONFIG_PATH).
- staging-deploy's session-log job installs engine dev dependencies: tsx is
  a devDependency at its pinned engine SHA, so --omit=dev left it missing.
- Workflow tests also check step if: parity, REPO_ROOT, the root checkout
  and other PATH-based engine launch forms.

+++ lib/schema.ts
@@ -42,12 +42,13 @@ export const devAgentConfigSchema = z.object({
    * read from the repo's own `.nvmrc`, `.node-version`, `.tool-versions` or
    * `package.json`, and only when none declares one does the engine's Node
    * apply. A default here would override what the repo already declares. An
-   * unquoted YAML number such as `node: 22` is accepted and kept as `"22"`.
+   * unquoted YAML integer such as `node: 22` is accepted and kept as `"22"`;
+   * a decimal must be quoted, since YAML reads `node: 22.10` as the number 22.1.
    */
   runtime: z
     .object({
       node: z
-        .union([z.string().min(1), z.number()])
+        .union([z.string().min(1), z.number().int()])
         .transform((value) => String(value))
         .optional(),
     })
